requirements.txt
dnspython==2.6.1
pymongo==4.7.1

aiogram~=3.7.0
fastapi~=0.111.0
cryptography~=42.0.8
PyYAML~=6.0.1
PyJWT~=2.8.0
uvicorn~=0.30.1
pydantic~=2.7.4
marshmallow~=3.21.3
motor~=3.5.1
tqdm~=4.66.4
colorlog~=6.8.2